[BLOOM-100] 로그아웃 및 RefershToken을 통한 토큰 재발급 구현

+++ src/main/kotlin/dnd11th/blooming/api/dto/user/SocialLoginResponse.kt
@@ -13,11 +13,11 @@ sealed class SocialLoginResponse {
         @field:Schema(description = "accessToken", example = "accessToken")
         val accessToken: String,
         @field:Schema(description = "accessToken 만료일자(초)", example = "604800(1주일)")
-        val expiresIn: Int,
+        val expiresIn: Long,
         @field:Schema(description = "refreshToken", example = "refreshToken")
         val refreshToken: String,
         @field:Schema(description = "refreshToken 만료일자(초)", example = "15552000(6개월)")
-        val refreshTokenExpiresIn: Int,
+        val refreshTokenExpiresIn: Long,
     ) : SocialLoginResponse()
 
     @Schema(
